fix(bench): third review round, and stop the 500-rejection bug being writable
The rejection trap is now closed rather than fixed a third time. `answerIo` in
`servers/io/respond.ts` reads and answers in one call that cannot reject and
cannot leave the response open, and both raw `node:http` handlers go through it.
That bug was written twice - once in `node-http.ts`, then again in
`servers/drivers/node.ts` after the first was fixed - and the next Node subject
would have been the third. The frameworks need none of this: Express 5, Fastify,
Hono and Nest all turn a rejected handler promise into a 5xx themselves.

`DriverUnit` and `DriversReport` move to `types.ts`, beside `ValidationReport` and
`LoggingReport`. `drivers-tables.ts` was keeping a hand-copied restatement of the
shape `drivers.ts` writes, with no compiler checking one against the other,
because the producer is an entrypoint with top-level effects and cannot be
imported from. The two `Report` types already there are the precedent.

`footprintRows(model)` ran twice per render of the benchmark page, once for the
length check and once for the prop.

Three other findings in this round were already fixed in 37bf430: the `broken`
filter, the batched seed insert, and the unread mean RSS.
